+++ jmetalsp-problem/src/main/java/org/uma/jmetalsp/problem/fda/FDA3.java
@@ -19,11 +19,11 @@ public class FDA3 extends FDA implements Serializable {
   private final int limitSupI = 1;
   private final int limitInfII = 1;
 
-  public FDA3(Observable<SingleObservedData<Double>> observable){
+  public FDA3(Observable<SingleObservedData<Integer>> observable){
     this(30,2, observable);
   }
 
-  public FDA3 (Integer numberOfVariables, Integer numberOfObjectives, Observable<SingleObservedData<Double>> observable) throws JMetalException {
+  public FDA3 (Integer numberOfVariables, Integer numberOfObjectives, Observable<SingleObservedData<Integer>> observable) throws JMetalException {
     super(observable) ;
     setNumberOfVariables(numberOfVariables);
     setNumberOfObjectives(numberOfObjectives);
